Add pyproject.yaml and reorg py files.
