+++ pyproject.toml
@@ -1,0 +1,29 @@
+[build-system]
+requires = ["setuptools>=45", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "pairx"
+dynamic = ["version"]
+description = " Pairwise mAtching of Intermediate Representations for eXplainability (PAIR-X)."
+readme = "README.md"
+requires-python = ">=3.10"
+dependencies = [
+    "torch>=2.0.0",
+    "torchvision>=0.15.0",
+    "transformers>=4.20.0",
+    "numpy>=1.21.0",
+    "pandas>=1.3.0",
+    "pillow>=8.0.0",
+    "opencv-python>=4.5.0",
+    "matplotlib>=3.5.0",
+    "zennit>=0.5.1",
+    "timm>=0.9.0"
+]
+
+[tool.setuptools.packages.find]
+where = ["src"]
+include = ["pairx*"]
+
+
+
