+++ src/pairx/xai_canonizers/__init__.py
@@ -1,0 +1,8 @@
+"""Canonizers for XAI methods"""
+
+from .efficientnet import EfficientNetBNCanonizer, EfficientNetCanonizer
+
+__all__ = [
+    "EfficientNetBNCanonizer",
+    "EfficientNetCanonizer",
+]
